Refactored logic to move out of the overrides out of the engine
